# Conflicts:
#	Assets/01.Script/0.Core/Manager/UpgradeManager.cs
#	Assets/02.Asset/Boss/SpawnBoss/Particles/FireBurstLight.prefab
#	Assets/06.ScriptAble/1.Core/PoolData.asset
#	UserSettings/EditorUserSettings.asset
#	UserSettings/Layouts/CurrentMaximizeLayout.dwlt
#	UserSettings/Layouts/default-2021.dwlt

+++ Assets/01.Script/2.Jaeby/UI/DamagePopup.cs
@@ -5,7 +5,7 @@ using DG.Tweening;
 
 public static class DamagePopup
 {
-    public static void PopupDamage(Vector3 targetPos, int damage)
+    public static void PopupDamage(Vector3 targetPos, float damage)
     {
         PopupPoolObject popupPoolObj = PoolManager.Instance.Pop(PoolType.PopupText) as PopupPoolObject;
         popupPoolObj.PopupTextCritical(targetPos + Vector3.up * 0.5f, damage.ToString());
